requirements.txt
beautifulsoup4==4.13.4
jsonschema==4.25.1
pydantic==2.11.7
PyYAML==6.0.2
requests==2.32.4
